+++ src/ai_commands.js
@@ -9,6 +9,7 @@ import { findFilepath, getIncludeText, functionDefinitionRegex } from './util';
 import conf from './ai_config';
 import { commandsList as _commandsList, commandsPrefix } from './commandsList';
 import { showInformationMessage, showErrorMessage, messages } from './ai_showMessage';
+import debugRemove from './commands/debugRemove';
 
 const { config } = conf;
 const aiOutCommon = window.createOutputChannel('AutoIt (global)', 'vscode-autoit-output');
@@ -1173,4 +1174,5 @@ export {
   openInclude,
   insertHeader,
   restartScript,
+  debugRemove,
 };